Confirmation button is now inside SchedulerCalendar component
