+++ src/components/SchedulerCalendarModal/SchedulerCalendarModal.js
@@ -1,6 +1,5 @@
 import React, { useContext } from "react";
 import SchedulerCalendar from "../../components/SchedulerCalendar/SchedulerCalendar";
-import Button from "@material-ui/core/Button";
 import { makeStyles } from "@material-ui/core/styles";
 import Modal from "@material-ui/core/Modal";
 import Backdrop from "@material-ui/core/Backdrop";
@@ -26,35 +25,13 @@ const useStyles = makeStyles(theme => ({
   h3: {
     fontSize: 20
   },
-  btn: {
-    background: "#A41F35",
-    borderRadius: 3,
-    border: 0,
-    color: "white",
-    height: 48,
-    padding: "0 100px",
-    margin: "0 auto",
-    display: "block",
-    marginTop: 20,
-    marginBottom: 20
-  },
-  "btn:hover": {
-    backgroundColor: "blue",
-    borderColor: "red",
-    boxShadow: "none"
-  },
   scheduler: { padding: "100px" }
 }));
 
 export default function SchedulerCalendarModal(props) {
   const classes = useStyles();
-  const { open, setOpen, setOpenConfirmation } = useContext(
-    SchedulePageContext
-  );
-  const handleClick = () => {
-    setOpen(false);
-    setOpenConfirmation(true);
-  };
+  const { open, setOpen } = useContext(SchedulePageContext);
+
   return (
     <div>
       <Modal
@@ -77,13 +54,6 @@ export default function SchedulerCalendarModal(props) {
               </p>
             </div>
             <SchedulerCalendar className={classes.scheduler} />
-            <Button
-              className={classes.btn}
-              variant="contained"
-              onClick={handleClick}
-            >
-              Confirm{" "}
-            </Button>
           </div>
         </Fade>
       </Modal>
